Fix updating JWKS service accounts
If we try to update a JWKS service account with the same
Issuer/Subject, we get an error that the Issuer/Subject must be unique
across all service accounts. I think this is a bug in TLSPC
(Issuer/Subject are still unique if we're acting on the same service
account).

Workaround this by only providing the Issuer/Subject to the update
payload if they have changed.