Don't set empty values in the state unless they have changed
If values are optional and we set them to the empty-value, terraform
interprets this as a change and attempts an update.

As we never set these value in the state (as they are optional and not
provided), this means that terraform attempts and update every single
time.

Fix by comparing values, and only updating the state if they are
different.