Add resource to validate GCP CloudProvider connections
Interface for this is a bit clumsy. In order for the cloudprovider to be
usable, we must issue an API call to validate the cloudprovider connection.
We can only do this after we've done all of the other things to put the
identity configuration in place. That means that we can't do it as part
of the create call for tlspc_cloudprovider_gcp (as we must take
parameters from that in order to set up the workload identity pool
provider).

However, this is effectively a write-once API. We can only ever validate
the connection, we can't unset that status once we've set it. That's not
a great mapping for terraform, but I think it's the best we can do.