Add CyberArk's Code of Conduct
I'm inclined to just use CyberArk's Code of Conduct as it stands.
The reality of it is that we are a part of CyberArk, and although it
might be a little confusing for consumers, for ancilliary files like
this, it's probably better just to keep them as-is, rather than maintain
sets of email aliases.